Fix blink transaction construction
The blink priority value was wrong, and so the required burn amount
wasn't getting added to the transaction data (and thus was getting
rejected by blink quorums).

This commit fixes it; it also slightly changes the logic to match
loki-core: we have "slow" priority, and anything else is blink.  (In
practice this shouldn't matter since only the two priorities--slow and
blink--weere exposed, but just makes the logic consistent).